requirements.txt
runpod
websocket-client
boto3
